requirements.txt
abstract==2022.7.10
asgiref==3.5.2
attrs==22.2.0
base32hex==1.0.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
chronometry==2020.11.12
colorama==0.4.6
colouration==2020.7.26
contourpy==1.0.6
cycler==0.11.0
dill==0.3.6
disk==2021.6.29
distlib==0.3.6
dj-shop-cart==5.2.0
Django==4.1.6
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
docker-pycreds==0.4.0
einops==0.5.0
english==2020.7.0
env==0.1.0
filelock==3.8.2
fonttools==4.38.0
func-timeout==4.3.5
fuzzywuzzy==0.18.0
geographiclib==1.52
geopy==2.2.0
google==3.0.0
graphviz==0.20.1
html5lib==1.1
idna==3.4
imageio==2.22.3
internet==2022.4.20.3
isodate==0.6.1
joblib==1.2.0
kiwisolver==1.4.4
laptop==1.0.0
lxml==4.9.1
map==1.3.0
matplotlib==3.6.2
mechanize==0.4.8
networkx==2.8.8
nmap==0.0.1
numpy==1.23.4
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.1
pathtools==0.1.2
pensieve==2022.4.20.1
Pillow==9.3.0
platformdirs==2.6.0
promise==2.3
protobuf==4.21.9
psutil==5.9.3
py4j==0.10.9.5
pyparsing==3.0.9
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2017.3
PyWavelets==1.4.1
PyYAML==6.0
ravenclaw==2020.11.12
rdflib==6.2.0
requests==2.28.1
robot==20071211
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
sentry-sdk==1.10.1
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
slytherin==2020.11.12
smmap==5.0.0
soupsieve==2.3.2.post1
soupspoon==2022.4.20
SPARQLWrapper==2.0.0
sqlparse==0.4.2
tech==0.0.0.20200724
threadpoolctl==3.1.0
tifffile==2022.10.10
toposort==1.7
torch==1.13.0
txt==2020.11.12
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.17.1
webencodings==0.5.1
well==0.1
